PS3 storage devices: use type unsigned long for region flags

+++ drivers/block/ps3disk.c
@@ -86,11 +86,11 @@ static const struct block_device_operations ps3disk_fops = {
 	.owner		= THIS_MODULE,
 };
 
-static unsigned int region_flags[] =
+static unsigned long region_flags[] =
 {
 	0x2, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0,
 };
-module_param_array(region_flags, uint, NULL, S_IRUGO);
+module_param_array(region_flags, ulong, NULL, S_IRUGO);
 MODULE_PARM_DESC(region_flags, "Region flags");
